requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
psutil==5.9.8
vdf==3.4
urllib3==2.2.0
pywin32==306
pywebview==4.4.1
werkzeug==3.0.1
